Extract the radio settings editor into web/js/ui/settings-panel.js
The settings tree, its per-value validation, tab rendering and the runtime
load/merge path now live in one module that owns that state privately.
Callers reach it through getGroups/setGroups/replaceState instead of
mutating a shared radioSettingsState object.

Cross-module calls (updateSerialActionState, setEditorView) go through a
late-bound actions registry, so no import cycle is introduced. Load-token
and runtime-API guards moved to state.js. No user-visible behavior change.
